# More efficient parsing of `ArraySeq`